+++ js/jquery.terminal.js
@@ -4,7 +4,7 @@
  *  __ / // // // // // _  // _// // / / // _  // _//     // //  \/ // _ \/ /
  * /  / // // // // // ___// / / // / / // ___// / / / / // // /\  // // / /__
  * \___//____ \\___//____//_/ _\_  / /_//____//_/ /_/ /_//_//_/ /_/ \__\_\___/
- *           \/              /____/                              version 2.23.0
+ *           \/              /____/                              version DEV
  *
  * This file is part of jQuery Terminal. https://terminal.jcubic.pl
  *
@@ -41,7 +41,7 @@
  *
  * broken image by Sophia Bai from the Noun Project (CC-BY)
  *
- * Date: Sun, 18 Apr 2021 11:24:37 +0000
+ * Date: Wed, 21 Apr 2021 11:19:19 +0000
  */
 /* global define, Map */
 /* eslint-disable */
@@ -4780,8 +4780,8 @@
     }
     // -------------------------------------------------------------------------
     $.terminal = {
-        version: '2.23.0',
-        date: 'Sun, 18 Apr 2021 11:24:37 +0000',
+        version: 'DEV',
+        date: 'Wed, 21 Apr 2021 11:19:19 +0000',
         // colors from https://www.w3.org/wiki/CSS/Properties/color/keywords
         color_names: [
             'transparent', 'currentcolor', 'black', 'silver', 'gray', 'white',
@@ -9411,11 +9411,10 @@
                         // Firefox won't reflow the cursor automatically, so
                         // hide it briefly then reshow it
                         cmd_cursor.hide();
+                        setTimeout(function() {
+                            cmd_cursor.show();
+                        }, 0);
                     }, 0);
-                    setTimeout(function() {
-                        // reshow cursor for firefox
-                        cmd_cursor.show();
-                    }, 10 /* this didn't work with 5ms */);
                     if ((settings.scrollOnEcho && options.scroll) || bottom) {
                         self.scroll_to_bottom();
                     }