+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.23.1
 ### Bugfix
 * fix reversed class in cmd (e.g. in less command)
+* fix hack to reflow the cursor in Firefox from #654
 
 ## 2.23.0
 ### Features